Added system for functions to free passed bstrings
Added new type freed_bstring to indicate that a function will free it's
arguments.
Function bautofree() can be used to indicate that you permit the
specified bstring to be freed by the function (there are no implicit
conversions between bstring and freed_bstring for a good reason!)
Function bautodestroy() can be used to free the bstring associated with
a freed_bstring and should be used inside functions that take
freed_bstring as arguments.
Function bautocpy() can be used to copy a bstring before passing it into
a function in case you want to keep your copy of it (it copies and
passes the copy into the function, where the passed copy will be freed).