More fixes for the Mac OS X build.
Dedicated cmake function added for OS X dependencies which handles a string split when finding SDL libraries as well as linking frameworks with cmake.

cpath is now set to dirname instead of dirname being called without storing the return because of different implementations of dirname which don't alter the path parameter.

emulator has a special define to redefine main function as SDL_main to avoid emulator crashing on startup once SDL_SetVideoMode is called.

testaux in bstring's main method was renamed to prevent it from linking as the main for emulator, deleting the file could also work as well however or removing the file from being included to be compiled but I just did this as a simple fix.

Added SDL and SDL dependencies for OS X in the CMakeLists for libtcod so that emulator will link with SDL properly to call SDLmain.

Removed the CustomSDLMain methods defined for Mac as they aren't defined in SDLmain anymore.