Fixing literally the weirdest Windows bug in existance.
It appears that the fseek/ftell functions on Windows have a "hole"
around 14880 upon which if you query for the location either by
ftell() or seeking with SEEK_CUR, you will be given ~14856.  When
reading character-by-character in the bfile library, it'd loop forever
since it would never pass this point.